large fix on lrc.c

+++ src/lobject.h
@@ -732,7 +732,7 @@ typedef struct Table {
 (check_exp((size&(size-1))==0, (cast_int((s) & ((size)-1)))))
 
 
-#define twoto(x)	(1<<(x))
+#define twoto(x)	(((size_t)1)<<(x))
 #define sizenode(t)	(twoto((t)->lsizenode))
